move ffmetabo code to featurefinder
