+++ src/openms/include/OpenMS/PROCESSING/DATAREDUCTION/sources.cmake
@@ -4,11 +4,7 @@ set(directory include/OpenMS/PROCESSING/DATAREDUCTION)
 ### list all header files of the directory here
 set(sources_list_h
 DataFilters.h
-Deisotoper.h
-ElutionPeakDetection.h
-FeatureFindingMetabo.h
 FeatureOverlapFilter.h
-MassTraceDetection.h
 SplineInterpolatedPeaks.h
 SplinePackage.h
 )
